SilkwormCoreSettings
