+++ cmd/common/settings.hpp
@@ -1,0 +1,34 @@
+/*
+   Copyright 2023 The Silkworm Authors
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+#pragma once
+
+#include <silkworm/infra/common/log.hpp>
+#include <silkworm/infra/rpc/server/server_config.hpp>
+#include <silkworm/node/common/settings.hpp>
+#include <silkworm/node/snapshot/settings.hpp>
+
+namespace silkworm::cmd::common {
+
+//! The overall settings for Silkworm Core component
+struct SilkwormCoreSettings {
+    silkworm::log::Settings log_settings;
+    silkworm::NodeSettings node_settings;
+    silkworm::rpc::ServerConfig server_settings;
+    SnapshotSettings snapshot_settings;
+};
+
+}  // namespace silkworm::cmd::common
